Internship/Chat Application/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
pydantic==2.6.4
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
cryptography==42.0.5
python-multipart==0.0.9

